Update project for OpenXR and Quest.
The project is now configured for OpenXR tracking/input and has the build settings appropriate for Quest. Also the readme is updated with environment/setup info if needed.  A working build of the scaffold app is now present and will be kept up-to-date in the "LatestBuild" directory of this repo.